feat: Support continuous learning.
1. Change the rule of extracting the feature of IP pattern.
2. Change the rule of extracting the feature of domain pattern.
3. Use continuous learning with RF classifier to reduce FPs and FNs.

+++ cli.py
@@ -44,6 +44,8 @@ def add_mode(dir: str):
     Args:
         dir: Folder path.
     """
+    if not os.path.exists(dir):
+        return
     for dirpath, dirnames, filenames in os.walk(dir):
         for dirname in dirnames:
             dir_path = os.path.join(dirpath, dirname)
@@ -54,17 +56,21 @@ def add_mode(dir: str):
             if not os.access(file_path, os.R_OK | os.W_OK):
                 os.chmod(file_path, 0o666)
 
-def decompress_packages(dataset_path: str) -> str:
+def decompress_packages(dataset_path: str, use_cache: bool = False) -> str:
     """Decompress packages.
     
     Args:
         dataset_path: Path of dataset.
+        use_cache: Use cache or not.
 
     Returns:
         Path of decompressed dataset.
     """
-    temp_dataset_path = os.path.abspath(f'.decompressed-packages-dataset')
+    temp_base_path = os.path.abspath(f'.decompressed-packages')
+    temp_dataset_path = os.path.abspath(os.path.join(temp_base_path, os.path.basename(dataset_path)))
 
+    if use_cache and os.path.exists(temp_dataset_path):
+        return temp_dataset_path
     if os.path.exists(temp_dataset_path):
         try:
             shutil.rmtree(temp_dataset_path)
@@ -101,49 +107,55 @@ def decompress_packages(dataset_path: str) -> str:
 
 def extract_cli():
     """Extract features from given dataset."""
-    dataset_name = args.dataset
-    dataset_path = os.path.join(SETTINGS['path']['datasets'], dataset_name)
-    dataset_path = os.path.abspath(decompress_packages(dataset_path))
-    if not os.path.exists(dataset_path):
-        print(f'Error: Dataset path {dataset_path} not found!')
-        exit(1)
-    feature_path = os.path.abspath(os.path.join(SETTINGS['path']['features'], dataset_name))
-    feature_position_path = os.path.abspath(os.path.join(SETTINGS['path']['feature-positions'], dataset_name))
+    dataset_names = args.dataset
+    use_cache = args.cache
+    for dataset_name in dataset_names:
+        dataset_path = os.path.join(SETTINGS['path']['datasets'], dataset_name)
+        dataset_path = os.path.abspath(decompress_packages(dataset_path, use_cache))
+        if not os.path.exists(dataset_path):
+            print(f'Error: Dataset path {dataset_path} not found!')
+            exit(1)
+        feature_path = os.path.abspath(os.path.join(SETTINGS['path']['features'], dataset_name))
+        feature_position_path = os.path.abspath(os.path.join(SETTINGS['path']['feature-positions'], dataset_name))
 
-    if os.path.exists(feature_path):
+        if os.path.exists(feature_path):
+            try:
+                shutil.rmtree(feature_path)
+            except PermissionError:
+                print(f'Error: Delete feature folder {feature_path} failed.')
+                traceback.print_exc()
+        os.makedirs(feature_path)
+
+        if os.path.exists(feature_position_path):
+            try:
+                shutil.rmtree(feature_position_path)
+            except PermissionError:
+                print(f'Error: Delete feature position folder {feature_path} failed.')
+                traceback.print_exc()
+        os.makedirs(feature_position_path)
+
         try:
-            shutil.rmtree(feature_path)
-        except PermissionError:
-            print(f'Error: Delete feature folder {feature_path} failed.')
+            cwd = os.getcwd()
+            os.chdir('feature-extract')
+            os.system(f'npm run start -- -d {dataset_path} {feature_path} {feature_position_path}')
+            os.chdir(cwd)
+        except Exception:
+            print(f'Error: Extract feature of dataset {dataset_name} failed.')
             traceback.print_exc()
-    os.makedirs(feature_path)
-
-    if os.path.exists(feature_position_path):
-        try:
-            shutil.rmtree(feature_position_path)
-        except PermissionError:
-            print(f'Error: Delete feature position folder {feature_path} failed.')
-            traceback.print_exc()
-    os.makedirs(feature_position_path)
-
-    try:
-        cwd = os.getcwd()
-        os.chdir('feature-extract')
-        os.system(f'npm run start -- -d {dataset_path} {feature_path} {feature_position_path}')
-        os.chdir(cwd)
-    except Exception:
-        print(f'Error: Extract feature of dataset {dataset_name} failed.')
-        traceback.print_exc()
 
 def train_cli():
     """Train model with given dataset."""
-    malicious_dataset_name = args.malicious
-    benign_dataset_name = args.benign
+    malicious_dataset_names = args.malicious
+    benign_dataset_names = args.benign
     model_name = args.model
     preprocess_method = args.preprocess
     action_name = args.action
-    malicous_csv_dir_path = os.path.join(SETTINGS['path']['features'], malicious_dataset_name)
-    benign_csv_dir_path = os.path.join(SETTINGS['path']['features'], benign_dataset_name)
+    malicous_csv_dir_paths = []
+    benign_csv_dir_paths = []
+    for malicious_dataset_name in malicious_dataset_names:
+        malicous_csv_dir_paths.append(os.path.join(SETTINGS['path']['features'], malicious_dataset_name))
+    for benign_dataset_name in benign_dataset_names:
+        benign_csv_dir_paths.append(os.path.join(SETTINGS['path']['features'], benign_dataset_name))
     hyperparameters = {}
 
     if preprocess_method == 'none':
@@ -180,30 +192,31 @@ def train_cli():
         hyperparameters['number_of_decision_trees'] = args.hyper_trees
         hyperparameters['maxium_depth'] = args.hyper_depth
 
-    train(malicous_csv_dir_path, benign_csv_dir_path, preprocess, model, action, hyperparameters)
+    train(malicous_csv_dir_paths, benign_csv_dir_paths, preprocess, model, action, hyperparameters)
 
 def predict_cli():
     """Predict packages."""
-    dataset_name = args.dataset
+    dataset_names = args.dataset
     model_name = args.model
 
-    report_name = f'{dataset_name}-{model_name}-report.csv'
-    report_content = 'package name, predict\n'
-    csv_dir_path = os.path.join(SETTINGS['path']['features'], dataset_name)
-    for feature_file_name in os.listdir(csv_dir_path):
-        feature_file_path = os.path.join(csv_dir_path, feature_file_name)
-        if model_name == 'MLP':
-            result = predict_package_MLP(feature_file_path)
-        elif model_name == 'NB':
-            result = predict_package_NB(feature_file_path)
-        elif model_name == 'SVM':
-            result = predict_package_SVM(feature_file_path)
-        elif model_name == 'RF':
-            result = predict_package_RF(feature_file_path)
-        report_content += feature_file_name[:-4] + ', ' + result + '\n'
+    for dataset_name in dataset_names:
+        report_name = f'{dataset_name}-{model_name}-report.csv'
+        report_content = 'package name, predict\n'
+        csv_dir_path = os.path.join(SETTINGS['path']['features'], dataset_name)
+        for feature_file_name in os.listdir(csv_dir_path):
+            feature_file_path = os.path.join(csv_dir_path, feature_file_name)
+            if model_name == 'MLP':
+                result = predict_package_MLP(feature_file_path)
+            elif model_name == 'NB':
+                result = predict_package_NB(feature_file_path)
+            elif model_name == 'SVM':
+                result = predict_package_SVM(feature_file_path)
+            elif model_name == 'RF':
+                result = predict_package_RF(feature_file_path)
+            report_content += feature_file_name[:-4] + ', ' + result + '\n'
 
-    with open(os.path.join(SETTINGS['path']['reports'], report_name), 'w') as f:
-        f.write(report_content)
+        with open(os.path.join(SETTINGS['path']['reports'], report_name), 'w') as f:
+            f.write(report_content)
 
 def predict_single_package(package_path: str):
     """Extract features and predict from given path."""
@@ -266,12 +279,13 @@ if __name__ == '__main__':
 
     # extract CLI parameters
     parser_extract = subparsers.add_parser('extract', help='extract features', description='Extract features from given dataset.')
-    parser_extract.add_argument('-d', '--dataset', type=str, required=True, help='dataset name', choices=DATASET_NAMES)
+    parser_extract.add_argument('-d', '--dataset', type=str, required=True, help='dataset name', choices=DATASET_NAMES, nargs='+')
+    parser_extract.add_argument('-c', '--cache', type=bool, help='use cache or not', default=False)
 
     # train CLI parameters
     parser_train = subparsers.add_parser('train', help='train model', description='Train model with given dataset.')
-    parser_train.add_argument('-m', '--malicious', type=str, required=True, help='malicious dataset name', choices=FEATURE_NAMES)
-    parser_train.add_argument('-b', '--benign', type=str, required=True, help='benign dataset name', choices=FEATURE_NAMES)
+    parser_train.add_argument('-m', '--malicious', type=str, required=True, help='malicious dataset name', choices=FEATURE_NAMES, nargs='+')
+    parser_train.add_argument('-b', '--benign', type=str, required=True, help='benign dataset name', choices=FEATURE_NAMES, nargs='+')
     parser_train.add_argument('-o', '--model', type=str, required=True, help='model name', choices=MODEL_NAMES)
     parser_train.add_argument('-p', '--preprocess', type=str, required=True, help='preprocess method', choices=PREPROCESS_METHOD_NAMES)
     parser_train.add_argument('-a', '--action', type=str, required=True, help='action', choices=['training', 'save'])
@@ -297,7 +311,7 @@ if __name__ == '__main__':
     # predict CLI parameters
     parser_predict = subparsers.add_parser('predict', help='predict package', description='Predict package with given model.')
     parser_predict.add_argument('-o', '--model', type=str, required=True, help='model name', choices=MODEL_NAMES)
-    parser_predict.add_argument('-d', '--dataset', type=str, help='dataset name', choices=FEATURE_NAMES)
+    parser_predict.add_argument('-d', '--dataset', type=str, help='dataset name', choices=FEATURE_NAMES, nargs='+')
     parser_predict.add_argument('-p', '--package-path', type=str, help='absolute package path')
 
     args = parser.parse_args()
